backend/requirements.txt
# Flask core
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Admin==1.6.1
Flask-CORS==4.0.0
Flask-WTF==1.2.1
Werkzeug==3.0.1

# Serialization
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Database
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9

# Async tasks
celery[redis]==5.3.6
redis==5.0.1

# Server
gunicorn==21.2.0

# ML/NLP (unchanged from Django)
sentence-transformers>=3.0.0
numpy==1.26.3
rank-bm25==0.2.2

# LLM providers
groq==0.4.2
httpx==0.27.0

# Document processing
pypdf2==3.0.1
python-docx==0.8.11
beautifulsoup4==4.12.3
lxml==5.1.0
requests==2.31.0

# Monitoring
sentry-sdk[flask]==1.39.1

# Utils
python-decouple==3.8
